Create caching Dark Sky client (closes #73)

+++ src/main/scala/com.snowplowanalytics/weather/package.scala
@@ -16,8 +16,8 @@ import hammock.Uri
 
 package object weather {
 
-  type Timestamp = Int
-  type Day       = Int // 0:00:00 timestamp of day
+  type Timestamp = Long
+  type Day       = Long // 0:00:00 timestamp of day
 
   trait WeatherRequest {
     def constructQuery(baseUri: Uri, apiKey: String): Uri
